DEL SeoulightCtrl,Model & Merge UserModel
