+++ routes/ServiceAPI.js
@@ -4,7 +4,6 @@ const imageUtil = require('../ImageUtil');
 
 const authCtrl = require('../controllers/AuthCtrl');
 const userCtrl = require('../controllers/UserCtrl');
-const seoulightCtrl = require('../controllers/SeoulightCtrl');
 const doboSTLECtrl = require('../controllers/DoboSLTECtrl');
 
 module.exports = (router) => {
@@ -23,7 +22,7 @@ module.exports = (router) => {
     .put(authCtrl.auth, imageUtil.uploadSingle, userCtrl.editAvatar);
 
   router.route('/seoulight/register') //시민해설사 신청
-    .post(authCtrl.auth, seoulightCtrl.reqSeoulight);
+    .post(authCtrl.auth, userCtrl.reqSeoulight);
 
   router.route('/users/feedback') //건의사항 
   .post(authCtrl.auth,userCtrl.addFeedback);
